timestamp extracted
